+++ client/apps/game/src/three/scenes/hexagon-scene.ts
@@ -11,7 +11,6 @@ import { HEX_SIZE, biomeModelPaths } from "@/three/scenes/constants";
 import { SystemManager } from "@/three/systems/system-manager";
 import { type SceneName } from "@/types";
 import { GRAPHICS_SETTING, GraphicsSettings, IS_FLAT_MODE } from "@/ui/config";
-import { getWorldPositionForHex } from "@/ui/utils/utils";
 import { type HexPosition } from "@bibliothecadao/eternum";
 import { LeftView, RightView, useUIStore, type AppStore, type SetupResult } from "@bibliothecadao/react";
 import gsap from "gsap";
@@ -19,6 +18,7 @@ import throttle from "lodash/throttle";
 import * as THREE from "three";
 import { type MapControls } from "three/examples/jsm/controls/MapControls";
 import { env } from "../../../env";
+import { getWorldPositionForHex } from "../utils";
 export abstract class HexagonScene {
   protected scene!: THREE.Scene;
   protected camera!: THREE.PerspectiveCamera;
